Admin role
Admin logged view added
