+++ dreambank/src/main/resources/templates/admin/adminMain.html
@@ -1,27 +1,93 @@
 <!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-
+<html
+        lang="en"
+        xmlns:th="http://www.thymeleaf.org"
+        xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity5"
+>
 <head>
-    <meta charset="ISO-8859-1">
-    <title>Customer List</title>
-
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
+    <title>Dream Bank</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <!-- fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Indie+Flower&family=Send+Flowers&display=swap" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Indie+Flower&family=Orbitron:wght@600&family=Send+Flowers&display=swap" rel="stylesheet">
+    <!-- bootstrap -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
-
+    <!-- css -->
     <link rel="stylesheet" href="/css/backgrounds.css">
 </head>
 <body>
 
-<div class="container my-2">
-    <h1 class="text-center">Admin Page</h1>
+<!-- nav bar -->
+<nav class="navbar navbar-expand-sm bg-primary navbar-dark">
+    <div class="container-fluid">
+        <ul class="navbar-nav">
+            <li class="nav-item">
+                <!-- Brand Logo -->
+                <a class="navbar-brand" href="/main">
+                    <img src="/image/brandlogo.png" alt="..." height="50">
+                </a>
+            </li>
+            <!-- Menu Start -->
+            <li class="nav-item">
+                <a class="nav-link" style="font-size: 25px;" href="/user/account">API</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" style="font-size: 25px;" href="/user/balance">Customer List</a>
+            </li>
+<!--            <li class="nav-item">-->
+<!--                <a class="nav-link" style="font-size: 25px;" href="/user/activities"></a>-->
+<!--            </li>-->
+            <!-- Menu End -->
+        </ul>
 
-    <a th:href="@{/admin/customers}" class="btn btn-primary btn-sm mb-3"> Customer List </a>
-<!--    <a th:href="@{/logout}" class="btn btn-primary btn-sm mb-3"> Log out </a>-->
-<!--    <a th:href="@{/logout}" class="btn btn-primary btn-sm mb-3"> Log out </a>-->
+        <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
+            <!-- Logout Button -->
+            <a href="/logout"><button type="button" class="btn btn-danger"> Logout </button></a>
+            <!-- GitHub Button -->
+            <li class="nav-item">
+                <a href="https://github.com/philippark89" id="giticon">
+                    <button type="button" class="btn btn-success">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-github" viewBox="0 0 16 16">
+                            <path d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27.68 0 1.36.09 2 .27 1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.012 8.012 0 0 0 16 8c0-4.42-3.58-8-8-8z"></path>
+                        </svg>
+                        GitHub
+                    </button>
+                </a>
+            </li>
+        </ul>
+    </div>
+</nav>
 
-    <br>
-    <a th:href="@{/logout}" class="btn btn-danger btn-sm mb-3"> Log out </a>
+<!-- main container -->
+<div class="container">
+    <div class="content text-center" style="height:200px">
+        <!--        <h1 class="text-warning" style="font-family: 'Indie Flower', cursive;" >Welcome To Dream Bank</h1>-->
+        <!--        <h2 th:text="${email}"></h2>-->
+        <h1 sec:authorize="hasRole('USER')">
+            You are logged in as Admin.<br>
+            What do you need to do?
+        </h1>
+    </div>
 </div>
+<!--    </div>-->
+
+<!-- script input to go to other pages -->
+<!--    <script>-->
+<!--      function go() {-->
+<!--        let page = document.getElementById("goInput").value.toLowerCase();-->
+<!--        if (page == "animals") {-->
+<!--          location.href = "/pages/animals.html";-->
+<!--        } else if (page == "tables") {-->
+<!--          location.href = "/pages/tables.html";-->
+<!--        } else {-->
+<!--          alert('type animals or tables');-->
+<!--        }-->
+<!--      }-->
+<!--    </script>-->
+
 </body>
 </html>
