+++ dreambank/src/main/resources/templates/login.html
@@ -30,11 +30,9 @@
 <!--        login / logout message handler-->
         <div th:if="${param.error}">
             <div class="text-center mt-4 name text-danger">Invalid username and password.</div>
-            <a href="/"><p class="text-center mt-2" >Go to Main</p></a>
         </div>
         <div th:if="${param.logout}">
             <div class="text-center mt-4 name text-success">You have been logged out.</div>
-            <a href="/"><p class="text-center mt-2" >Go to Main</p></a>
         </div>
     </div>
 </body>
